application table is now a component of application
