+++ src/app/applications/components/table.component.ts
@@ -1,0 +1,235 @@
+import { ApplicationRaw, FilterStringOperator, SessionTaskOptionEnumField, TaskOptionEnumField } from '@aneoconsultingfr/armonik.api.angular';
+import { CdkDragDrop, DragDropModule, moveItemInArray } from '@angular/cdk/drag-drop';
+import { NgFor, NgIf } from '@angular/common';
+import { AfterViewInit, Component, EventEmitter, Input, OnInit, Output, ViewChild, inject } from '@angular/core';
+import { MatButtonModule } from '@angular/material/button';
+import { MatDialog } from '@angular/material/dialog';
+import { MatIconModule } from '@angular/material/icon';
+import { MatMenuModule } from '@angular/material/menu';
+import { MatPaginator, MatPaginatorModule } from '@angular/material/paginator';
+import { MatSort, MatSortModule } from '@angular/material/sort';
+import { MatTableModule } from '@angular/material/table';
+import { RouterModule } from '@angular/router';
+import { TaskSummaryFiltersOr } from '@app/tasks/types';
+import { TaskStatusColored, ViewTasksByStatusDialogData } from '@app/types/dialog';
+import { CountTasksByStatusComponent } from '@components/count-tasks-by-status.component';
+import { FiltersToolbarComponent } from '@components/filters/filters-toolbar.component';
+import { TableEmptyDataComponent } from '@components/table/table-empty-data.component';
+import { TableActionsToolbarComponent } from '@components/table-actions-toolbar.component';
+import { TableContainerComponent } from '@components/table-container.component';
+import { ViewTasksByStatusDialogComponent } from '@components/view-tasks-by-status-dialog.component';
+import { EmptyCellPipe } from '@pipes/empty-cell.pipe';
+import { IconsService } from '@services/icons.service';
+import { TasksByStatusService } from '@services/tasks-by-status.service';
+import { ApplicationsIndexService } from '../services/applications-index.service';
+import { ApplicationRawColumnKey, ApplicationRawFieldKey, ApplicationRawListOptions } from '../types';
+
+@Component({
+  selector: 'app-application-table',
+  standalone: true,
+  template: `
+<app-table-container>
+  <table mat-table matSort [matSortActive]="options.sort.active" recycleRows matSortDisableClear [matSortDirection]="options.sort.direction" [dataSource]="data" cdkDropList cdkDropListOrientation="horizontal" [cdkDropListDisabled]="lockColumns" (cdkDropListDropped)="onDrop($event)">
+
+    <ng-container *ngFor="let column of displayedColumns" [matColumnDef]="column">
+      <!-- Header -->
+      <th mat-header-cell mat-sort-header [disabled]="isNotSortableColumn(column)" *matHeaderCellDef cdkDrag appNoWrap>
+        {{ columnToLabel(column) }}
+      </th>
+      <!-- Application Column -->
+      <ng-container *ngIf="isSimpleColumn(column)">
+        <td mat-cell *matCellDef="let element" appNoWrap>
+          {{ element[column] | emptyCell }}
+        </td>
+      </ng-container>
+      <!-- Application's Tasks Count by Status -->
+      <ng-container *ngIf="isCountColumn(column)">
+        <td mat-cell *matCellDef="let element" appNoWrap>
+        <app-count-tasks-by-status
+          [statuses]="tasksStatusesColored"
+          [queryParams]="createTasksByStatusQueryParams(element.name, element.version)"
+          [filters]="countTasksByStatusFilters(element.name, element.version)"
+        ></app-count-tasks-by-status>
+        </td>
+      </ng-container>
+      <!-- Action -->
+      <ng-container *ngIf="isActionsColumn(column)">
+        <td mat-cell *matCellDef="let element" appNoWrap>
+         <button mat-icon-button [matMenuTriggerFor]="menu"
+          aria-label="Actions" i18n-aria-label>
+            <mat-icon [fontIcon]="getIcon('more')"></mat-icon>
+         </button>
+         <mat-menu #menu="matMenu">
+            <a mat-menu-item [routerLink]="['/sessions']" [queryParams]="createViewSessionsQueryParams(element.name, element.version)">
+              <mat-icon aria-hidden="true" [fontIcon]="getIcon('view')"></mat-icon>
+              <span i18n>See session</span>
+            </a>
+         </mat-menu>
+        </td>
+      </ng-container>
+    </ng-container>
+
+    <!-- Empty -->
+    <tr *matNoDataRow>
+      <td [attr.colspan]="displayedColumns.length">
+        <app-table-empty-data></app-table-empty-data>
+      </td>
+    </tr>
+
+    <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
+    <tr mat-row *matRowDef="let row; columns: displayedColumns;"></tr>
+  </table>
+
+  <mat-paginator [length]="total" [pageIndex]="options.pageIndex" [pageSize]="options.pageSize" [pageSizeOptions]="[5, 10, 25, 100]" aria-label="Select page of applications">
+  </mat-paginator>
+</app-table-container>
+  `,
+  styles: [
+    
+  ],
+  providers: [
+    ApplicationsIndexService,
+    TasksByStatusService,
+    MatDialog,
+    IconsService
+  ],
+  imports: [
+    TableActionsToolbarComponent,
+    FiltersToolbarComponent,
+    TableContainerComponent,
+    MatPaginatorModule,
+    TableEmptyDataComponent,
+    MatMenuModule,
+    CountTasksByStatusComponent,
+    MatSortModule,
+    NgFor,
+    NgIf,
+    MatTableModule,
+    MatIconModule,
+    RouterModule,
+    EmptyCellPipe,
+    DragDropModule,
+    MatButtonModule
+  ]
+})
+export class ApplicationTableComponent implements OnInit, AfterViewInit {
+
+  @Input({required: true}) displayedColumns: ApplicationRawColumnKey[] = [];
+  @Input({required: true}) options: ApplicationRawListOptions;
+  @Input({required: true}) data: ApplicationRaw.AsObject[] = [];
+  @Input({required: true}) total: number;
+  @Input() lockColumns = false;
+
+  @Output() optionsChange = new EventEmitter<never>();
+
+  tasksStatusesColored: TaskStatusColored[] = [];
+  @ViewChild(MatSort) sort: MatSort;
+  @ViewChild(MatPaginator) paginator: MatPaginator;
+  
+  readonly _applicationsIndexService = inject(ApplicationsIndexService);
+  readonly _tasksByStatusService = inject(TasksByStatusService);
+  readonly _dialog = inject(MatDialog);
+  readonly _iconsService = inject(IconsService);
+
+  ngOnInit(): void {
+    this.tasksStatusesColored = this._tasksByStatusService.restoreStatuses('applications');
+    this.total = this.data.length;
+  }
+
+  ngAfterViewInit(): void {
+    this.sort.sortChange.subscribe(() => {
+      this.options.pageIndex = 0; // If the user change the sort order, reset back to the first page.
+      this.options.sort = {
+        active: this.sort.active as ApplicationRawFieldKey,
+        direction: this.sort.direction
+      };
+      this.optionsChange.emit();
+    });
+
+    this.paginator.page.subscribe(() => {
+      this.options.pageIndex = this.paginator.pageIndex;
+      this.options.pageSize = this.paginator.pageSize;
+      this.optionsChange.emit();
+    });
+  }
+
+  getIcon(name: string): string {
+    return this._iconsService.getIcon(name);
+  }
+
+  columnToLabel(column: ApplicationRawColumnKey): string {
+    return this._applicationsIndexService.columnToLabel(column);
+  }
+
+  isActionsColumn(column: ApplicationRawColumnKey): boolean {
+    return this._applicationsIndexService.isActionsColumn(column);
+  }
+
+  isCountColumn(column: ApplicationRawColumnKey): boolean {
+    return this._applicationsIndexService.isCountColumn(column);
+  }
+
+  isSimpleColumn(column: ApplicationRawColumnKey): boolean {
+    return this._applicationsIndexService.isSimpleColumn(column);
+  }
+
+  isNotSortableColumn(column: ApplicationRawColumnKey): boolean {
+    return this._applicationsIndexService.isNotSortableColumn(column);
+  }
+
+  createTasksByStatusQueryParams(name: string, version: string) {
+    return {
+      [`0-options-${TaskOptionEnumField.TASK_OPTION_ENUM_FIELD_APPLICATION_NAME}-${FilterStringOperator.FILTER_STRING_OPERATOR_EQUAL}`]: name,
+      [`0-options-${TaskOptionEnumField.TASK_OPTION_ENUM_FIELD_APPLICATION_VERSION}-${FilterStringOperator.FILTER_STRING_OPERATOR_EQUAL}`]: version,
+    };
+  }
+
+  createViewSessionsQueryParams(name: string, version: string) {
+    return {
+      [`0-options-${SessionTaskOptionEnumField.TASK_OPTION_ENUM_FIELD_APPLICATION_NAME}-${FilterStringOperator.FILTER_STRING_OPERATOR_EQUAL}`]: name,
+      [`0-options-${SessionTaskOptionEnumField.TASK_OPTION_ENUM_FIELD_APPLICATION_VERSION}-${FilterStringOperator.FILTER_STRING_OPERATOR_EQUAL}`]: version,
+    };
+  }
+
+  personalizeTasksByStatus() {
+    const dialogRef = this._dialog.open<ViewTasksByStatusDialogComponent, ViewTasksByStatusDialogData>(ViewTasksByStatusDialogComponent, {
+      data: {
+        statusesCounts: this.tasksStatusesColored,
+      }
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (!result) {
+        return;
+      }
+
+      this.tasksStatusesColored = result;
+      this._tasksByStatusService.saveStatuses('applications', result);
+    });
+  }
+
+  countTasksByStatusFilters(applicationName: string, applicationVersion: string): TaskSummaryFiltersOr {
+    return [
+      [
+        {
+          for: 'options',
+          field: TaskOptionEnumField.TASK_OPTION_ENUM_FIELD_APPLICATION_NAME,
+          value: applicationName,
+          operator: FilterStringOperator.FILTER_STRING_OPERATOR_EQUAL
+        },
+        {
+          for: 'options',
+          field: TaskOptionEnumField.TASK_OPTION_ENUM_FIELD_APPLICATION_VERSION,
+          value: applicationVersion,
+          operator: FilterStringOperator.FILTER_STRING_OPERATOR_EQUAL
+        }
+      ]
+    ];
+  }
+
+  onDrop(event: CdkDragDrop<string[]>) {
+    moveItemInArray(this.displayedColumns, event.previousIndex, event.currentIndex);
+
+    this._applicationsIndexService.saveColumns(this.displayedColumns);
+  }
+}
