+++ src/app/applications/index.component.spec.ts
@@ -1,12 +1,6 @@
-import { TaskStatus } from '@aneoconsultingfr/armonik.api.angular';
-import { CdkDragDrop } from '@angular/cdk/drag-drop';
 import { TestBed } from '@angular/core/testing';
-import { MatDialog } from '@angular/material/dialog';
-import { MatPaginator, PageEvent } from '@angular/material/paginator';
-import { MatSort, Sort } from '@angular/material/sort';
 import { BehaviorSubject, Subject, throwError } from 'rxjs';
 import { DATA_FILTERS_SERVICE } from '@app/tokens/filters.token';
-import { TaskStatusColored } from '@app/types/dialog';
 import { DataFilterService } from '@app/types/filter-definition';
 import { FiltersOr } from '@app/types/filters';
 import { AutoRefreshService } from '@services/auto-refresh.service';
@@ -14,7 +8,6 @@ import { FiltersService } from '@services/filters.service';
 import { IconsService } from '@services/icons.service';
 import { NotificationService } from '@services/notification.service';
 import { ShareUrlService } from '@services/share-url.service';
-import { TasksByStatusService } from '@services/tasks-by-status.service';
 import { IndexComponent } from './index.component';
 import { ApplicationsGrpcService } from './services/applications-grpc.service';
 import { ApplicationsIndexService } from './services/applications-index.service';
@@ -99,11 +92,6 @@ describe('Application component', () => {
     resetFilters: jest.fn(),
   };
 
-  const mockTasksByStatusService = {
-    restoreStatuses: jest.fn(),
-    saveStatuses: jest.fn()
-  };
-
   const mockNotificationService = {
     error: jest.fn()
   };
@@ -117,25 +105,11 @@ describe('Application component', () => {
     createInterval: jest.fn(() => intervalRefreshSubject),
   };
 
-  let dialogSubject: BehaviorSubject<TaskStatusColored[] | undefined>;
-
   beforeEach(() => {
     component = TestBed.configureTestingModule({
       providers: [
         IndexComponent,
-        {provide: TasksByStatusService, useValue: mockTasksByStatusService },
         {provide: NotificationService, useValue: mockNotificationService },
-        {provide: MatDialog, useValue:
-          {
-            open: () => {
-              return {
-                afterClosed: () => {
-                  return dialogSubject;
-                }
-              };
-            }
-          }
-        },
         IconsService,
         FiltersService,
         DataFilterService,
@@ -146,21 +120,9 @@ describe('Application component', () => {
         { provide: AutoRefreshService, useValue: mockAutoRefreshService },
       ]
     }).inject(IndexComponent);
-
-    component.paginator = {
-      pageIndex: 1,
-      pageSize: 25,
-      page: new Subject<PageEvent>()
-    } as unknown as MatPaginator;
-
-    component.sort = new MatSort();
-    component.sort.active = 'sessions';
-    component.sort.direction = 'asc';
     
     component.ngOnInit();
     component.ngAfterViewInit();
-
-    component.displayedColumns = ['name', 'namespace', 'service', 'version', 'actions', 'count'];
   });
   
   it('Should run', () => {
@@ -173,7 +135,6 @@ describe('Application component', () => {
     expect(mockApplicationsFilterService.restoreFilters).toHaveBeenCalled();
     expect(mockApplicationIndexService.restoreIntervalValue).toHaveBeenCalled();
     expect(mockShareUrlService.generateSharableURL).toHaveBeenCalledWith(component.options, component.filters);
-    expect(mockTasksByStatusService.restoreStatuses).toHaveBeenCalledWith('applications');
     expect(component.availableColumns).toBe(mockApplicationIndexService.availableColumns);
   });
 
@@ -184,18 +145,13 @@ describe('Application component', () => {
 
     describe('on sort change', () => {
 
-      const sort: Sort = {
-        active: 'id',
-        direction: 'asc'
-      };
-
       it('should reset pageIndex', () => {
-        component.sort.sortChange.next(sort);
-        expect(component.paginator.pageIndex).toEqual(0);
+        component.sort.next();
+        expect(component.options.pageIndex).toEqual(0);
       });
 
       it('should load data', () => {
-        component.sort.sortChange.next(sort);
+        component.sort.next();
         expect(mockGrpcApplicationsService.list$).toHaveBeenCalledWith(options, sampleFiltersOr);
         expect(component.total).toEqual(3);
         expect(component.data).toEqual(dataSample.applications);
@@ -204,12 +160,7 @@ describe('Application component', () => {
   });
 
   it('should load data on page change', () => {
-    component.paginator.page.next({
-      pageIndex: 2,
-      previousPageIndex: 1,
-      pageSize: 25,
-      length: 10
-    });
+    component.paginator.next();
     expect(mockGrpcApplicationsService.list$).toHaveBeenCalledWith(options, sampleFiltersOr);
     expect(component.total).toEqual(3);
     expect(component.data).toEqual(dataSample.applications);
@@ -247,40 +198,9 @@ describe('Application component', () => {
   it('should return column labels', () => {
     expect(component.columnsLabels()).toEqual(mockApplicationIndexService.columnsLabels);
   });
-  
-  it('should return the label of a column', () => {
-    component.columnToLabel('name');
-    expect(mockApplicationIndexService.columnToLabel).toHaveBeenCalledWith('name');
-  });
-
-  it('should check if a column is an action', () => {
-    component.isActionsColumn('actions');
-    expect(mockApplicationIndexService.isActionsColumn).toHaveBeenCalledWith('actions');
-  });
-
-  it('should check if a column is a count column', () => {
-    component.isCountColumn('count');
-    expect(mockApplicationIndexService.isCountColumn).toHaveBeenCalledWith('count');
-  });
-
-  it('should check if a column is a simple column', () => {
-    component.isSimpleColumn('name');
-    expect(mockApplicationIndexService.isSimpleColumn).toHaveBeenCalledWith('name');
-  });
-
-  it('should check if a column is a sortable column', () => {
-    component.isNotSortableColumn('actions');
-    expect(mockApplicationIndexService.isNotSortableColumn).toHaveBeenCalledWith('actions');
-  });
 
   it('should get page icon', () => {
     expect(component.getPageIcon('applications')).toEqual('apps');
-  });
-
-  it('should get required icons', () => {
-    expect(component.getIcon('tune')).toEqual('tune');
-    expect(component.getIcon('more')).toEqual('more_vert');
-    expect(component.getIcon('view')).toEqual('visibility');
   });
 
   it('should refresh', () => {
@@ -333,7 +253,7 @@ describe('Application component', () => {
     component.onFiltersChange(newFilterOr);
     
     expect(mockApplicationsFilterService.saveFilters).toHaveBeenCalledWith(newFilterOr);
-    expect(component.paginator.pageIndex).toEqual(0);
+    expect(component.options.pageIndex).toEqual(0);
     expect(spyRefresh).toHaveBeenCalled();
     expect(component.filters).toEqual(newFilterOr);
   });
@@ -342,7 +262,7 @@ describe('Application component', () => {
     mockApplicationsFilterService.resetFilters.mockImplementationOnce(() => []);
     component.onFiltersReset();
     expect(mockApplicationsFilterService.resetFilters).toHaveBeenCalled();
-    expect(component.paginator.pageIndex).toEqual(0);
+    expect(component.options.pageIndex).toEqual(0);
     expect(component.filters).toEqual([]);
   });
 
@@ -352,74 +272,11 @@ describe('Application component', () => {
     expect(mockAutoRefreshService.autoRefreshTooltip).toHaveBeenCalledWith(13);
   });
 
-  it('should change column order', () => {
-    const event = {
-      previousIndex: 0,
-      currentIndex: 1
-    } as unknown as CdkDragDrop<string[]>;
-    component.onDrop(event);
-    expect(mockApplicationIndexService.saveColumns).toHaveBeenCalledWith(component.displayedColumns);
-    expect(component.displayedColumns).toEqual(['namespace', 'name', 'service', 'version', 'actions', 'count']);
-  });
-
   it('should stop the auto-refresh', () => {
     const spyStopRefresh = jest.spyOn(component.stopInterval, 'next');
     component.intervalValue = 0;
     component.handleAutoRefreshStart();
     expect(spyStopRefresh).toHaveBeenCalled();
-  });
-
-  it('should count tasks by status of the filters', () => {
-    expect(component.countTasksByStatusFilters('unified_api', '1.0.0'))
-      .toEqual([[
-        {
-          for: 'options',
-          field: 5,
-          value: 'unified_api',
-          operator: 0
-        },
-        {
-          for: 'options',
-          field: 6,
-          value: '1.0.0',
-          operator: 0
-        }
-      ]]);
-  });
-
-  it('should create the query params of the tasks by status', () => {
-    expect(component.createTasksByStatusQueryParams('name', 'version'))
-      .toEqual({
-        ['0-options-5-0']: 'name',
-        ['0-options-6-0']: 'version'
-      });
-  });
-
-  it('should create the query params to view sessions', () => {
-    expect(component.createViewSessionsQueryParams('name', 'version'))
-      .toEqual({
-        ['0-options-5-0']: 'name',
-        ['0-options-6-0']: 'version'
-      });
-  });
-
-  it('should permit to personalize tasks by status', () => {
-    dialogSubject = new BehaviorSubject<TaskStatusColored[] | undefined>([{
-      color: 'green',
-      status: TaskStatus.TASK_STATUS_COMPLETED
-    }]);
-    component.personalizeTasksByStatus();
-    expect(component.tasksStatusesColored).toEqual([{
-      color: 'green',
-      status: TaskStatus.TASK_STATUS_COMPLETED
-    }]);
-    expect(mockTasksByStatusService.saveStatuses).toHaveBeenCalled();
-  });
-
-  it('should not personalize if there is no result', () => {
-    dialogSubject = new BehaviorSubject<TaskStatusColored[] | undefined>(undefined);
-    component.personalizeTasksByStatus();
-    expect(mockTasksByStatusService.saveStatuses).toHaveBeenCalledTimes(0);
   });
 
   describe('onLockColumnsChange', () => {
